Use ncurses or pdcurses based on config
Created two configurations for dub to use which determines which version
of curses it is going to use.  The installation uses one of these two
commands:

  dub build --config=pdcurses (for PDCurses)
  dub build --config=ncurses  (for ncurses)

These two configurations need to be tested on compatible systems.

+++ src/display.d
@@ -32,7 +32,7 @@ void display_mon( monst m )
 
 void display_map_mons( map to_display )
 {
-  ulong d = to_display.m.length;
+  uint d = to_display.m.length;
   monst mn;
   foreach( c; 0 .. d )
   {
